관리자 페이지 유저 상세정보 불러오기 구현

+++ src/main/java/com/guestbook/dimention/repository/UserRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface UserRepository  extends JpaRepository<SiteUser, Long> {
+public interface UserRepository  extends JpaRepository<SiteUser, Integer> {
     Optional<SiteUser> findByuserId(String userId);
 }
